:sparkles: Add parseKeyTrail and parseMultipleArguments dependencies
parseKeyTrail accepts an object and a path that represents an object's property. It returns the value of the object's property at the specified path or undefined. This is useful because it prevents needing to check each node along the way: tile && tile.button && tile.button.text could be portrayed instead as parseKeyTrail('button.text', tile). parseMultipleArguments is used to accept a single argument, multiple arguments, or an array of arguments and return an array of arguments. The function that uses this should use ...rest and pass the result as args = parseMultipleArguments(rest). This is useful if functions should be flexible with their arguments, so you know args will always be an array of values