:sparkles: Create super sort function
Super Sort accepts one or more arguments consisting of strings or functions. The strings should represent properties to sort on, and the functions should return 1 or -1 if sorting conditions are met. Super Sort sorts an array by the provided conditions in the order they are provided. By default, each property is sorted in ascending order. prepending the property name with '<' or '>' will explicitly tell the function to sort the property in ascending or descending order respectively