:rotating_light: Allow reassignment of reduce accumulator
no-param-reassign does not allow for modifying properties of arguments passed into functions. This is an annoyance for reduce functions where it is a common pattern to modify an accumulator argument and return it. Instead of copying the accumulator and returning its copy, lets allow reassignment of properties for common accumulator names: 'acc', 'accumulator', 'object', 'result'