:recycle: Flip arguments for parseKeyTrail utility function
The original decision to start with the object key and then the target object was inspired by a philosophy that functions and objects should be the last argument, because if the function or object literal is hardcoded, arguments following the function or literal are confusing. Because of this, each time I've used this function, the arguments have felt backwards. It makes more sense to first pass the target object, then pass the path within the target object we're searching for. In addition, this is a function where we will always want to pass a reference to an object, because if you're creating the object literal as an argument for this function, you already know the result of this function.