:boom: Nuke all packaging code
Parcel broke the library by renaming all functions, even ones in the global scope to 20+ character gibberish names.

This is unworkable because the script cannot be used as a JavaScript file because none of the functions preserved their names.

I spent over an hour trying to put together a solution using Grunt and Babel to build the project instead, but Babel would not refactor my CommonJS modules into something like AMD no matter what I tried in the configuration.

This was unworkable because the paths of the CommonJS in the Babel output didn't point to anything. That default behavior is asinine.

The closest thing I could find to a correct solution would be to use the ESM format instead of CommonJS and then to install a Babel plugin that would convert ESM to AMD. This is promising and looks like it will work, but the Babel documentatino severely lacks examples, contexts, or links to the settings that it refers to and I am at the end of my rope right now.

Maybe there is better tooling for the extremely simple use-case of turning a handful of files into a single JavaScript file, but that is a problem to solve on another day.